+++ include/samurai/schemes/fv.hpp
@@ -11,12 +11,12 @@
 #include "fv/flux_based/explicit_flux_based_scheme.hpp"
 #include "fv/scheme_operators.hpp"
 
-#include "fv/operators/convection_FV__lin.hpp"
-#include "fv/operators/convection_FV__nonlin.hpp"
-#include "fv/operators/diffusion_FV.hpp"
-#include "fv/operators/diffusion_FV_old.hpp"
-#include "fv/operators/divergence_FV.hpp"
-#include "fv/operators/divergence_FV__nonlin.hpp"
-#include "fv/operators/gradient_FV.hpp"
-#include "fv/operators/identity_FV.hpp"
-#include "fv/operators/zero_operator_FV.hpp"
+#include "fv/operators/convection_lin.hpp"
+#include "fv/operators/convection_nonlin.hpp"
+#include "fv/operators/diffusion.hpp"
+#include "fv/operators/diffusion_old.hpp"
+#include "fv/operators/divergence.hpp"
+#include "fv/operators/flux_divergence.hpp"
+#include "fv/operators/gradient.hpp"
+#include "fv/operators/identity.hpp"
+#include "fv/operators/zero_operator.hpp"
