+++ include/samurai/schemes/fv/operators/diffusion.hpp
@@ -1,7 +1,7 @@
 #pragma once
 #include "../flux_based/flux_based_scheme__lin_het.hpp"
 #include "../flux_based/flux_based_scheme__lin_hom.hpp"
-#include "divergence_FV.hpp"
+#include "divergence.hpp"
 
 namespace samurai
 {
